# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FusionHub API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface ErrorEnvelope {
  error: string;
}

export interface App {
  id: number;
  userId: string;
  name: string;
  url: string;
  /** Emoji or image URL used as the app icon */
  icon: string;
  /** Accent color for the app card (hex) */
  color: string;
  isFavorite: boolean;
  position: number;
  category: string;
  launchCount: number;
  /** @nullable */
  lastLaunchedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateAppBody {
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string;
  /** @minLength 1 */
  url: string;
  icon?: string;
  color?: string;
  category?: string;
  isFavorite?: boolean;
}

export interface UpdateAppBody {
  /**
   * @minLength 1
   * @maxLength 100
   */
  name?: string;
  url?: string;
  icon?: string;
  color?: string;
  category?: string;
  isFavorite?: boolean;
  position?: number;
}

export type ReorderAppsBodyOrderItem = {
  id: number;
  position: number;
};

export interface ReorderAppsBody {
  order: ReorderAppsBodyOrderItem[];
}

export type AppStatsCategoryCountsItem = {
  category: string;
  count: number;
};

export interface AppStats {
  totalApps: number;
  favoriteApps: number;
  totalLaunches: number;
  categoryCounts: AppStatsCategoryCountsItem[];
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  /**
   * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
   */
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type ListAppsParams = {
  /**
   * Search apps by name or URL
   */
  search?: string;
  /**
   * Filter by category
   */
  category?: string;
};

export type ListRecentAppsParams = {
  limit?: number;
};

export type ReorderApps200 = {
  success: boolean;
};

export type DeleteApp200 = {
  success: boolean;
};