# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FusionHub API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get the currently authenticated user
 */
export const GetCurrentAuthUserHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

export const GetCurrentAuthUserResponse = zod.object({
  user: zod.union([
    zod.object({
      id: zod.string(),
      email: zod.string().email().nullable(),
      firstName: zod.string().nullable(),
      lastName: zod.string().nullable(),
      profileImageUrl: zod.string().nullable(),
    }),
    zod.null(),
  ]),
});

/**
 * @summary Start the browser OIDC login flow
 */
export const BeginBrowserLoginQueryParams = zod.object({
  returnTo: zod.coerce
    .string()
    .optional()
    .describe(
      "Relative path to redirect to after login (must start with `\/`). Defaults to `\/`.",
    ),
});

/**
 * @summary Complete the browser OIDC login flow
 */
export const HandleBrowserLoginCallbackQueryParams = zod.object({
  code: zod.coerce.string().optional(),
  state: zod.coerce.string().optional(),
  iss: zod.coerce.string().url().optional(),
});

/**
 * @summary Clear the session and begin OIDC logout
 */
export const LogoutBrowserSessionHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

/**
 * @summary List all apps for the current user
 */
export const ListAppsQueryParams = zod.object({
  search: zod.coerce.string().optional().describe("Search apps by name or URL"),
  category: zod.coerce.string().optional().describe("Filter by category"),
});

export const ListAppsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  name: zod.string(),
  url: zod.string(),
  icon: zod.string().describe("Emoji or image URL used as the app icon"),
  color: zod.string().describe("Accent color for the app card (hex)"),
  isFavorite: zod.boolean(),
  position: zod.number(),
  category: zod.string(),
  launchCount: zod.number(),
  lastLaunchedAt: zod.coerce.date().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListAppsResponse = zod.array(ListAppsResponseItem);

/**
 * @summary Create a new app
 */
export const createAppBodyNameMax = 100;

export const createAppBodyIconDefault = `🌐`;
export const createAppBodyColorDefault = `#f97316`;
export const createAppBodyCategoryDefault = `General`;
export const createAppBodyIsFavoriteDefault = false;

export const CreateAppBody = zod.object({
  name: zod.string().min(1).max(createAppBodyNameMax),
  url: zod.string().min(1),
  icon: zod.string().default(createAppBodyIconDefault),
  color: zod.string().default(createAppBodyColorDefault),
  category: zod.string().default(createAppBodyCategoryDefault),
  isFavorite: zod.boolean().default(createAppBodyIsFavoriteDefault),
});

/**
 * @summary Get usage statistics for current user's apps
 */
export const GetAppStatsResponse = zod.object({
  totalApps: zod.number(),
  favoriteApps: zod.number(),
  totalLaunches: zod.number(),
  categoryCounts: zod.array(
    zod.object({
      category: zod.string(),
      count: zod.number(),
    }),
  ),
});

/**
 * @summary List favorite apps for the current user
 */
export const ListFavoriteAppsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  name: zod.string(),
  url: zod.string(),
  icon: zod.string().describe("Emoji or image URL used as the app icon"),
  color: zod.string().describe("Accent color for the app card (hex)"),
  isFavorite: zod.boolean(),
  position: zod.number(),
  category: zod.string(),
  launchCount: zod.number(),
  lastLaunchedAt: zod.coerce.date().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListFavoriteAppsResponse = zod.array(ListFavoriteAppsResponseItem);

/**
 * @summary List recently launched apps
 */
export const listRecentAppsQueryLimitDefault = 5;

export const ListRecentAppsQueryParams = zod.object({
  limit: zod.coerce.number().default(listRecentAppsQueryLimitDefault),
});

export const ListRecentAppsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  name: zod.string(),
  url: zod.string(),
  icon: zod.string().describe("Emoji or image URL used as the app icon"),
  color: zod.string().describe("Accent color for the app card (hex)"),
  isFavorite: zod.boolean(),
  position: zod.number(),
  category: zod.string(),
  launchCount: zod.number(),
  lastLaunchedAt: zod.coerce.date().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListRecentAppsResponse = zod.array(ListRecentAppsResponseItem);

/**
 * @summary Reorder apps by updating their positions
 */
export const ReorderAppsBody = zod.object({
  order: zod.array(
    zod.object({
      id: zod.number(),
      position: zod.number(),
    }),
  ),
});

export const ReorderAppsResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get a specific app
 */
export const GetAppParams = zod.object({
  id: zod.coerce.number(),
});

export const GetAppResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  name: zod.string(),
  url: zod.string(),
  icon: zod.string().describe("Emoji or image URL used as the app icon"),
  color: zod.string().describe("Accent color for the app card (hex)"),
  isFavorite: zod.boolean(),
  position: zod.number(),
  category: zod.string(),
  launchCount: zod.number(),
  lastLaunchedAt: zod.coerce.date().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update an app
 */
export const UpdateAppParams = zod.object({
  id: zod.coerce.number(),
});

export const updateAppBodyNameMax = 100;

export const UpdateAppBody = zod.object({
  name: zod.string().min(1).max(updateAppBodyNameMax).optional(),
  url: zod.string().optional(),
  icon: zod.string().optional(),
  color: zod.string().optional(),
  category: zod.string().optional(),
  isFavorite: zod.boolean().optional(),
  position: zod.number().optional(),
});

export const UpdateAppResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  name: zod.string(),
  url: zod.string(),
  icon: zod.string().describe("Emoji or image URL used as the app icon"),
  color: zod.string().describe("Accent color for the app card (hex)"),
  isFavorite: zod.boolean(),
  position: zod.number(),
  category: zod.string(),
  launchCount: zod.number(),
  lastLaunchedAt: zod.coerce.date().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete an app
 */
export const DeleteAppParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteAppResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Record an app launch (increments launch count)
 */
export const LaunchAppParams = zod.object({
  id: zod.coerce.number(),
});

export const LaunchAppResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  name: zod.string(),
  url: zod.string(),
  icon: zod.string().describe("Emoji or image URL used as the app icon"),
  color: zod.string().describe("Accent color for the app card (hex)"),
  isFavorite: zod.boolean(),
  position: zod.number(),
  category: zod.string(),
  launchCount: zod.number(),
  lastLaunchedAt: zod.coerce.date().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});