# lib/api-zod/src/generated/types/app.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FusionHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface App {
  id: number;
  userId: string;
  name: string;
  url: string;
  /** Emoji or image URL used as the app icon */
  icon: string;
  /** Accent color for the app card (hex) */
  color: string;
  isFavorite: boolean;
  position: number;
  category: string;
  launchCount: number;
  /** @nullable */
  lastLaunchedAt?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}